Split DTOs for voice channels to make them more readable.
