+++ API/DTOs/Voice/IceCandidateDto.cs
@@ -1,0 +1,8 @@
+﻿namespace API.DTOs.Voice;
+
+public class IceCandidateDto
+{
+    public required string Candidate { get; set; }
+    public string? SdpMid { get; set; }
+    public int? SdpMLineIndex { get; set; }
+}
