+++ API/DTOs/Voice/VoiceChannelPresenceDto.cs
@@ -1,0 +1,7 @@
+﻿namespace API.DTOs.Voice;
+
+public class VoiceChannelPresenceDto
+{
+    public required string ChannelId { get; set; }
+    public List<VoiceParticipantDto> Participants { get; set; } = [];
+}
