+++ API/DTOs/Voice/VoiceChannelPresenceSnapshotDto.cs
@@ -1,0 +1,7 @@
+﻿namespace API.DTOs.Voice;
+
+public class VoiceChannelPresenceSnapshotDto
+{
+    public required string ChatRoomId { get; set; }
+    public List<VoiceChannelPresenceDto> Channels { get; set; } = [];
+}
