+++ API/DTOs/Voice/VoiceIceCandidateMessageDto.cs
@@ -1,0 +1,8 @@
+﻿namespace API.DTOs.Voice;
+
+public class VoiceIceCandidateMessageDto
+{
+    public required string ChannelId { get; set; }
+    public required string FromConnectionId { get; set; }
+    public required IceCandidateDto Candidate { get; set; }
+}
