+++ API/DTOs/Voice/VoicePeerUpdateDto.cs
@@ -1,0 +1,7 @@
+﻿namespace API.DTOs.Voice;
+
+public class VoicePeerUpdateDto
+{
+    public required string ChannelId { get; set; }
+    public required VoiceParticipantDto Participant { get; set; }
+}
