+++ API/DTOs/Voice/VoiceSignalMessageDto.cs
@@ -1,0 +1,8 @@
+﻿namespace API.DTOs.Voice;
+
+public class VoiceSignalMessageDto
+{
+    public required string ChannelId { get; set; }
+    public required string FromConnectionId { get; set; }
+    public required SessionDescriptionDto Description { get; set; }
+}
