+++ API/Program.cs
@@ -1,23 +1,26 @@
-using API.Middleware;
-using API.SignalR;
 using System.Net;
+using API.Middleware;
+using API.Services;
+using API.SignalR;
+using API.SignalR.EventHandlers;
+using Application.ChatAppearances.Validators;
+using Application.ChatRoomRoles.Validators;
 using Application.ChatRooms.Queries;
 using Application.ChatRooms.Validators;
-using Application.ChatAppearances.Validators;
 using Application.Core;
-using Application.EmojiPreferences.Validators;
 using Application.Development;
+using Application.EmojiPreferences.Validators;
 using Application.Friends.Validators;
 using Application.Interfaces;
 using Application.Profiles.Validators;
 using Application.Status.Validators;
 using Domain;
+using Domain.Enums;
 using FluentValidation;
 using Infrastructure.Email;
 using Infrastructure.Media;
 using Infrastructure.Security;
 using Infrastructure.Services;
-using API.Services;
 using Infrastructure.Storage;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
@@ -25,11 +28,8 @@ using Microsoft.AspNetCore.Mvc.Authorization;
 using Microsoft.EntityFrameworkCore;
 using Minio;
 using Persistance;
-using Resend;
-using Application.ChatRoomRoles.Validators;
-using Domain.Enums;
-using API.SignalR.EventHandlers;
 using Persistance.Security;
+using Resend;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -55,13 +55,15 @@ builder.Services.AddHttpClient<ILinkPreviewService, LinkPreviewService>(client =
     {
         AllowAutoRedirect = true,
         AutomaticDecompression = DecompressionMethods.All,
-        CookieContainer = new CookieContainer(),
+        CookieContainer = new CookieContainer()
     };
 
-    handler.CookieContainer.Add(new Uri("https://www.youtube.com"), new Cookie("CONSENT", "YES+1", "/", ".youtube.com"));
+    handler.CookieContainer.Add(new Uri("https://www.youtube.com"),
+        new Cookie("CONSENT", "YES+1", "/", ".youtube.com"));
     handler.CookieContainer.Add(new Uri("https://youtube.com"), new Cookie("CONSENT", "YES+1", "/", ".youtube.com"));
     handler.CookieContainer.Add(new Uri("https://youtu.be"), new Cookie("CONSENT", "YES+1", "/", ".youtu.be"));
-    handler.CookieContainer.Add(new Uri("https://youtube-nocookie.com"), new Cookie("CONSENT", "YES+1", "/", ".youtube-nocookie.com"));
+    handler.CookieContainer.Add(new Uri("https://youtube-nocookie.com"),
+        new Cookie("CONSENT", "YES+1", "/", ".youtube-nocookie.com"));
     handler.CookieContainer.Add(new Uri("https://google.com"), new Cookie("CONSENT", "YES+1", "/", ".google.com"));
 
     return handler;
@@ -73,10 +75,7 @@ builder.Services.AddMediatR(x =>
     x.AddOpenBehavior(typeof(ValidationBehavior<,>));
 });
 builder.Services.AddHttpClient<ResendClient>();
-builder.Services.Configure<ResendClientOptions>(opt =>
-{
-    opt.ApiToken = builder.Configuration["Resend:ApiToken"]!;
-});
+builder.Services.Configure<ResendClientOptions>(opt => { opt.ApiToken = builder.Configuration["Resend:ApiToken"]!; });
 builder.Services.AddTransient<IResend, ResendClient>();
 builder.Services.AddTransient<IEmailSender<User>, EmailSender>();
 builder.Services.AddScoped<IUserAccessor, UserAccessor>();
@@ -97,7 +96,7 @@ if (builder.Environment.IsDevelopment())
 }
 else
 {
-    // Azure Blob Storage for production / LEGACY / WE NOT USE MINIO FOR BOTH
+    // Azure Blob Storage for production / LEGACY / WE NOW USE MINIO FOR BOTH
     // builder.Services.AddSingleton(sp =>
     //     new BlobServiceClient(builder.Configuration.GetConnectionString("AzureStorage")));
     // builder.Services.AddScoped<IFileStorage, AzureBlobStorage>();
@@ -112,6 +111,7 @@ else
     );
     builder.Services.AddScoped<IFileStorage, MinioStorage>();
 }
+
 builder.Services.AddScoped<IFriendsNotificationService, FriendsNotificationService>();
 builder.Services.AddScoped<IUserStatusService, UserStatusService>();
 builder.Services.AddSingleton<IVoiceChannelPresenceService, VoiceChannelPresenceService>();
@@ -132,30 +132,22 @@ builder.Services.AddValidatorsFromAssemblyContaining<UpdateChatRoomRoleValidator
 builder.Services.AddTransient<ExceptionMiddleware>();
 builder.Services.AddHostedService<MediaCleanupService>();
 builder.Services.AddIdentityApiEndpoints<User>(opt =>
-{
-    opt.User.RequireUniqueEmail = true;
-    if (builder.Environment.IsProduction())
     {
-        opt.SignIn.RequireConfirmedEmail = true;
-    }
-})
-.AddRoles<IdentityRole>()
-.AddEntityFrameworkStores<AppDbContext>();
+        opt.User.RequireUniqueEmail = true;
+        if (builder.Environment.IsProduction()) opt.SignIn.RequireConfirmedEmail = true;
+    })
+    .AddRoles<IdentityRole>()
+    .AddEntityFrameworkStores<AppDbContext>();
 builder.Services.AddAuthorization(opt =>
 {
-    opt.AddPolicy(IsOwnerStrings.IsChatRoomOwner, policy =>
-    {
-        policy.Requirements.Add(new IsOwnerRequirement());
-    });
+    opt.AddPolicy(IsOwnerStrings.IsChatRoomOwner, policy => { policy.Requirements.Add(new IsOwnerRequirement()); });
 
     foreach (var permissionName in ChatRoomPermissions.All.Keys)
-    {
         opt.AddPolicy(permissionName, policy =>
         {
             policy.Requirements.Add(
                 new HasPermissionRequirement(permissionName));
         });
-    }
 });
 builder.Services.AddTransient<IAuthorizationHandler, IsOwnerRequirementHandler>();
 builder.Services.AddTransient<IAuthorizationHandler, HasPermissionRequirementHandler>();
@@ -163,7 +155,7 @@ MessageCrypto.Initialize(builder.Configuration);
 
 var clientAppOrigins = builder.Configuration["ClientAppUrl"]?
     .Split(';', StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries);
-var corsOrigins = (clientAppOrigins is { Length: > 0 })
+var corsOrigins = clientAppOrigins is { Length: > 0 }
     ? clientAppOrigins
     : ["https://localhost:3000"];
 
@@ -205,15 +197,12 @@ try
     await context.Database.MigrateAsync();
 
     if (builder.Environment.IsDevelopment())
-    {
         await DbInitializer.SeedData(context, userManager, rolePermissionService, chatRoomRoleService);
-    }
 }
 catch (Exception ex)
 {
     var logger = services.GetRequiredService<ILogger<Program>>();
     logger.LogError(ex, "An error occurred during migration");
 }
+
 app.Run();
-
-
